+++ SalCap/src/main/resources/templates/index.html
@@ -346,7 +346,7 @@ body {
     </main>
 
     <!-- JavaScript 연결 -->
-    <script src="salaryCalculator.js"></script>
+    <script type="module" src="salaryCalculator.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/xlsx/0.14.3/xlsx.full.min.js"></script> <!-- 엑셀 변환 -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/FileSaver.js/1.3.8/FileSaver.min.js"></script><!-- 파일 저장 -->
     <script src="https://cdn.jsdelivr.net/npm/xlsx-js-style@1.2.0/dist/xlsx.min.js"></script> <!-- 엑셀 스타일 적용 -->
@@ -354,4 +354,6 @@ body {
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jspdf/2.5.1/jspdf.umd.min.js"></script> <!-- pdf 생성-->
     <script src="https://unpkg.com/jspdf-autotable"></script> <!--동적으로 표 생성-->
     <script src="https://developers.kakao.com/sdk/js/kakao.min.js"></script>
+	<!--<script src="https://code.jquery.com/jquery-3.4.1.js"></script> //jQuery 사용 위해-->
+	
 </html>
